refactor: extract internal/logrotate for log cleanup
Move CleanupOldRuns out of internal/config into a focused logrotate
package. Config keeps Load/Validate/ResolveLogsDir; only the sort
import is dropped (os/time still needed).

Closes #68
